utils/pyproject.toml
[project]
name = "vand-utils"
version = "0.1.0"
description = "VAND 4.0 Challenge — shared helper utilities"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "tifffile>=2023.0",
    "Pillow>=9.4",
    "scikit-learn>=1.3",
    "tqdm>=4.64",
]

[tool.hatch.build.targets.wheel]
packages = ["retail", "industrial"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
